Rimozione delle varie ricerche
Rimozione delle varie ricerche; unica classe Ricerche con tutti i metodi di ricerca; aggiunta metodo ricerca per dimensione memoria.
